Config type generic, bundle source factory as a parameter
